Extract `renderHTML()` to general UI script and use in the modal-dialog content rendering.
